requirements.txt
# Dependencias para modelos de IA
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.7.0

# Dependencias de base de datos
sqlalchemy>=2.0.0

# Dependencias adicionales
pydantic>=2.0.0
python-dotenv>=1.0.0
pandas>=1.5.0
numpy>=1.24.0

# Para bases de datos externas (opcional)
# psycopg2-binary>=2.9.0  # PostgreSQL
# mysql-connector-python>=8.0.0  # MySQL

# Alternativas a ToolFront (implementación personalizada)
# Usaremos SQLAlchemy + modelos de IA directamente